Rename utils/utils.js to utils/common.js.
