+++ packages/ckeditor5-table/src/commands/insertrowcommand.js
@@ -9,7 +9,7 @@
 
 import Command from '@ckeditor/ckeditor5-core/src/command';
 import { findAncestor } from './utils';
-import { getRowIndexes, getSelectionAffectedTableCells } from '../utils/utils';
+import { getRowIndexes, getSelectionAffectedTableCells } from '../utils/common';
 
 /**
  * The insert row command.
