+++ packages/ckeditor5-table/src/commands/mergecellcommand.js
@@ -10,7 +10,7 @@
 import Command from '@ckeditor/ckeditor5-core/src/command';
 import TableWalker from '../tablewalker';
 import { isHeadingColumnCell, findAncestor } from './utils';
-import { getTableCellsContainingSelection } from '../utils/utils';
+import { getTableCellsContainingSelection } from '../utils/common';
 
 /**
  * The merge cell command.
