+++ packages/ckeditor5-table/src/utils/common.js
